Hotel rating added

+++ src/main/java/com/example/hotelbookingservice/dto/HotelResponse.java
@@ -15,7 +15,7 @@ public class HotelResponse {
 
     private Integer distance;
 
-    private short rating;
+    private Float rating;
 
     private Integer ratingsNum;
 }
